Re-write entire css file every macro expansion
While slower on initial compilation we need to re-write the css file to
make the figwheel workflow be correct. Appending the new css each change
is incorrect when we delete a selector as the old selector will cascade
down.

The theoretically optimal way to do this would be to append to file when
we are adding components but re-write the component when there are
changes. This would be performant for both initial compilation and small
changes. However, this is complicated and we should keep the simple and
correct thing until we feel the pain from a slow initial build.