OMCSS-12: Eval constants inside `style`
This patch introduces evaluation of constants inside the style forms
that Garden will generate CSS output from. It is now possible to write
components such as:

```clojure
(ns my-ns.core
  (:require [om-css.core :as oc :refer-macros [defui]]
            [om-css.dom :as dom]
            [my-ns.constants :as c]))

(defui Foo
   static oc/Style
   (style [_]
     [[:.root {:background-color "tomato"}]
      [:.section (merge c/some-style ;; <== notice this
                   {:background-color :green})]])
   Object
   (render [this]
     (dom/div {:class :root} "div with class :root"
       (dom/section {:class :section} "section with class :section"))))
```

There's a caveat to the inclusion of such constants that is a result of
the CSS being generated at compile time. As such, all the constants used
inside a style form must reside in a `.cljc` file.